Pull stock data for company over a period
Pulls stock data for company over a period of time using the Alpha Vantage API. The company used for the exercise is Tesla. 
With the stock data for a week and the sentiment from tweets about the company, the group can correlate the two to help an investor make decisions regarding the stock price of the company.